requirements.txt
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.35.0
peft>=0.8.0
scikit-learn>=1.3.0
tqdm
Pillow
